feat: add enterprise and organization octokit classes
